+++ backend/app/src/regs.c
@@ -31,8 +31,7 @@ void print_regs(pid_t pid) {
   printf("\n");
 }
 
-// TODO : long -> ull ?
-long get_rsp(pid_t pid) {
+unsigned long long get_rsp(pid_t pid) {
   struct user_regs_struct regs;
 
   x_ptrace_get_register_info(pid, &regs);
